refactor: Remove verifySignature
